refactor: assetBalance APIをreceiptから分離
- assetBalanceApiをfeatures/assetBalance配下へ移動
- useAssetBalanceとAssetBalanceページの参照先を新APIへ更新
- receiptApiからassetBalance依存を削除して責務を明確化
